--- src/server/__tests__/mock-tests.js
@@ -1,5 +1,0 @@
-import test from 'ava';
-
-test('init', t => {
-  t.is(1, 1);
-});
