+++ src/universal/redux/__tests__/AuthEngine-tests.js
@@ -1,5 +1,3 @@
-import test from 'ava';
-import sinon from 'sinon';
 import {applyMiddleware, createStore, combineReducers} from 'redux';
 import thunk from 'redux-thunk';
 import auth from '../authDuck';
@@ -12,67 +10,55 @@ const REDUCER_NAME = 'marvin'; // use a different key than default for testing
 const reducers = combineReducers({[REDUCER_NAME]: auth});
 const store = createStore(reducers, {}, applyMiddleware(thunk));
 
-test.before(() => {
-  // stub cashay.query to return mocked data:
-  const cashayQueryUserStub = () => {
-    return ({
-      data: {
-        user: {
-          avatar: null,
-          email: null,
-          id: null,
-          name: null
-        }
+const cashayQueryUserStub = () => {
+  return ({
+    data: {
+      user: {
+        avatar: null,
+        email: null,
+        id: null,
+        name: null
       }
-    });
-  };
-  sinon.stub(cashay, 'query').callsFake(cashayQueryUserStub);
-});
+    }
+  });
+};
+cashay.query = jest.fn(cashayQueryUserStub);
 
-test.after.always('cleanup', () => {
-  cashay.query.restore();
-});
-
-test('constructor stores store', t => {
+test('constructor stores store', () => {
   const rae = new ReduxAuthEngine(store, REDUCER_NAME);
-  t.deepEqual(rae.store, store);
+  expect(rae.store).toEqual(store);
 });
 
-test('saveToken saves token', t => {
-  t.plan(3);
-
+test('saveToken saves token', (done) => {
   const cb = (unused, token) => {
-    t.is(unused, null);
-    t.is(token, testToken);
+    expect(unused).toBe(null);
+    expect(token).toBe(testToken);
+    done();
   };
 
   const rae = new ReduxAuthEngine(store, REDUCER_NAME);
   rae.saveToken('aName', testToken, null, cb);
-  t.deepEqual(
-    rae.store.getState(),
-    {
+  expect(rae.store.getState())
+    .toEqual({
       [REDUCER_NAME]: {
         obj: testTokenData,
         token: testToken
       }
-    }
-  );
+    });
 });
 
-test('removeToken removes token', t => {
-  t.plan(3);
-
+test('removeToken removes token', (done) => {
   const cb = (unused, token) => {
-    t.is(unused, null);
-    t.is(token, testToken);
+    expect(unused).toBe(null);
+    expect(token).toBe(testToken);
+    done();
   };
 
   const rae = new ReduxAuthEngine(store, REDUCER_NAME);
   rae.saveToken('aName', testToken);
   rae.removeToken('aName', cb);
-  t.deepEqual(
-    rae.store.getState(),
-    {
+  expect(rae.store.getState())
+    .toEqual({
       [REDUCER_NAME]: {
         token: null,
         obj: {
@@ -83,16 +69,14 @@ test('removeToken removes token', t => {
           sub: null
         }
       }
-    }
-  );
+    });
 });
 
-test('loadToken callsback with token', t => {
-  t.plan(2);
-
+test('loadToken callsback with token', (done) => {
   const cb = (unused, token) => {
-    t.is(unused, null);
-    t.is(token, testToken);
+    expect(unused).toBe(null);
+    expect(token).toBe(testToken);
+    done();
   };
 
   const rae = new ReduxAuthEngine(store, REDUCER_NAME);
