--- src/universal/utils/testEnv.babel.js
@@ -1,3 +1,0 @@
-
-const run = require('../../server/init').run;
-run();
