--- tests/pretest.babel.js
@@ -1,3 +1,0 @@
-require('babel-register');
-require('babel-polyfill');
-require('./pretest')();
